requirements.txt
fastapi
uvicorn
httpx
sqlalchemy
alembic
pydantic
pytest
python-dotenv
httpx